Move hit objects UI elements to `ui.v2`
