+++ src/ru/nsu/ccfit/zuev/osu/game/GameScene.java
@@ -30,9 +30,9 @@ import com.reco1l.andengine.Modifiers;
 import com.reco1l.andengine.Anchor;
 import com.reco1l.andengine.sprite.VideoSprite;
 import com.reco1l.andengine.ExtendedScene;
-import com.reco1l.osu.hitobjects.FollowPointConnection;
+import com.osudroid.ui.v2.hitobjects.FollowPointConnection;
 import com.osudroid.ui.v2.hud.GameplayHUD;
-import com.reco1l.osu.hitobjects.SliderTickSprite;
+import com.osudroid.ui.v2.hitobjects.SliderTickSprite;
 import com.osudroid.ui.v2.hud.elements.HUDPPCounter;
 import com.osudroid.ui.v1.BlockAreaFragment;
 import com.reco1l.osu.multiplayer.Multiplayer;
